Add exception and logic.
