+++ src/lesson1/Course.java
@@ -1,3 +1,5 @@
+package lesson1;
+
 public class Course {
     ObstacleCourse[] courses;
