+++ src/lesson1/Main.java
@@ -1,3 +1,5 @@
+package lesson1;
+
 public class Main {
     public static void main(String[] args) {
         ObstacleCourse[] oc = {new Wall(3), new RunningTrack(7), new Wall(1)};
@@ -6,8 +8,8 @@ public class Main {
         c.doIt(team); // Просим команду пройти полосу
         team.showResultOfTeam(); // Показываем результаты
 
-//        Android xyz = new Android("Misha", 100, 12);
-//        RunningTrack road = new RunningTrack(100);
+//        lesson1.Android xyz = new lesson1.Android("Misha", 100, 12);
+//        lesson1.RunningTrack road = new lesson1.RunningTrack(100);
 //        road.doIt(xyz);
 //        xyz.result();
     }
