+++ src/lesson1/ObstacleCourse.java
@@ -1,3 +1,5 @@
+package lesson1;
+
 public abstract class ObstacleCourse {
     public abstract void doIt(Skills runner);
 }
