+++ src/lesson1/RunningTrack.java
@@ -1,3 +1,5 @@
+package lesson1;
+
 public class RunningTrack extends ObstacleCourse {
     private int distance;
