+++ src/lesson1/Team.java
@@ -1,3 +1,5 @@
+package lesson1;
+
 public class Team {
     String teamName;
     Skills[] teamRunner;
